auto fetch vk file when deamon run
update ci
