+++ .circleci/config.yml
@@ -115,7 +115,7 @@ jobs:
 
   deps_linux:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     working_directory: /go/src/github.com/filecoin-project/venus
     resource_class: xlarge
     steps:
@@ -169,7 +169,7 @@ jobs:
 
   build_linux:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     working_directory: /go/src/github.com/filecoin-project/venus
     resource_class: xlarge
     parameters:
@@ -223,7 +223,7 @@ jobs:
 
   unit_test_linux:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     parallelism: 2 # Check .codecov.yml "after_n_builds" if changing this
     working_directory: /go/src/github.com/filecoin-project/venus
     resource_class: xlarge
@@ -255,7 +255,7 @@ jobs:
 
   integration_test_linux:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     parallelism: 2 # Check .codecov.yml "after_n_builds" if changing this
     working_directory: /go/src/github.com/filecoin-project/venus
     resource_class: xlarge
@@ -289,7 +289,7 @@ jobs:
 
   functional_test_linux:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     parallelism: 2
     working_directory: /go/src/github.com/filecoin-project/venus
     resource_class: xlarge
@@ -327,7 +327,7 @@ jobs:
 
   publish_release:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     resource_class: small
     parameters:
       <<: *nightly_param
@@ -353,7 +353,7 @@ jobs:
 
   build_docker_img:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     resource_class: xlarge
     parameters:
       <<: *nightly_param
@@ -443,7 +443,7 @@ jobs:
 
   trigger_nightly_devnet_deploy:
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     resource_class: small
     steps:
       - setup_remote_docker
@@ -479,7 +479,7 @@ jobs:
       <<: *user_devnet_param
       <<: *staging_devnet_param
     docker:
-      - image: circleci/golang:1.14.1-stretch
+      - image: circleci/golang:1.16.2-stretch
     resource_class: small
     steps:
       - checkout
