+++ cmd/main.go
@@ -164,12 +164,11 @@ var RootCmdDaemon = &cmds.Command{
 
 // all top level commands, not available to daemon
 var rootSubcmdsLocal = map[string]*cmds.Command{
-	"daemon":       daemonCmd,
-	"fetch":        fetchCmd,
-	"version":      versionCmd,
-	"leb128":       leb128Cmd,
-	"seed":         seedCmd,
-	"fetch-params": fetchParamCmd,
+	"daemon":  daemonCmd,
+	"fetch":   fetchCmd,
+	"version": versionCmd,
+	"leb128":  leb128Cmd,
+	"seed":    seedCmd,
 }
 
 // all top level commands, available on daemon. set during init() to avoid configuration loops.
